Move UPnP handling to NetManager, re-use existing net peers.
This means we get proper handling of logging and such.
